backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0